Rename module back to fanuc_motion_program_exec_client
