+++ examples/multi_coord_example.py
@@ -1,4 +1,4 @@
-from fanuc_motion_program_exec import *
+from fanuc_motion_program_exec_client import *
 
 tp_lead = TPMotionProgram()
 tp_follow = TPMotionProgram()
